// src/models/enums/ErrorEnum.ts
export enum ErrorEnum {
  BLUETOOTH_INIT_ERROR = "BLUETOOTH_INIT_ERROR",
  CONNECTION_ATTEMPT_ERROR = "CONNECTION_ATTEMPT_ERROR",
  DEVICE_NOT_CONNECTED = "DEVICE_NOT_CONNECTED",
  WIFI_NOT_IMPLEMENTED = "WIFI_NOT_IMPLEMENTED",
  DEVICE_HAS_DISCONNECTED = "DEVICE_HAS_DISCONNECTED",
  CONNECTING_FAILED = "CONNECTING_FAILED",
}