// src/models/enums/CommandResultEnum.ts
export enum CommandResultEnum {
  SUCCESS = 0,
  ERROR = 1,
  ACTION_NOT_NECESSARY = 2,
  I2C_ERROR = 3,
  LIDAR_ERROR = 4,
  LIDAR_RESOLUTION_ERROR = 5,
  LIDAR_RANGING_MODE_ERROR = 6,
  LIDAR_POWER_MODE_ERROR = 7,
  LIDAR_INTEGRATION_TIME_ERROR = 8,
  LIDAR_SHARPENER_ERROR = 9,
  LIDAR_TARGET_ORDER_ERROR = 10,
  LIDAR_RANGING_ERROR = 11,
  LIDAR_MEASUREMENT_ERROR = 12,
  LUNA_MEASUREMENT_ERROR = 13,
  NO_DETECTORS_CONFIGURED_ERROR = 14,
}